refactor: 💡 started splitting general home page to componnts
