+++ src/components/pages/GeneralHome/GeneralHome.style.js
@@ -43,50 +43,39 @@ const useStyles = makeStyles((theme) => ({
 		minHeight: '7.9vh',
 		overflow: 'hidden',
 	},
-	mostClickedIdeasWrapper: {
-		'display': 'flex',
-		'justifyContent': 'space-between',
-		'padding': '0 10px 0 10px',
-		'maxHeight': '9vh',
-		'minHeight': 72,
-		'overflowY': 'hidden',
-		'minWidth': '25.2vh',
-		'margin': '10px',
-		'marginTop': 8,
-		'alignItems': 'center',
-		'borderRadius': '8px',
-		'backgroundColor': '#EDF2FB',
-		'willChange': 'transform',
-		'animation': 'mostClickedIdeas 20s linear infinite',
-		'&:hover': {
-			cursor: 'pointer',
-		},
-	},
-	mostClickedIdeasTitle: {
-		color: '#8197ae',
-		fontSize: '.9rem',
-	},
-	mostClickedIdeasContent: {
-		color: '#151515',
-		fontSize: '1rem',
-	},
+	// mostClickedIdeasWrapper: {
+	// 	'display': 'flex',
+	// 	'justifyContent': 'space-between',
+	// 	'padding': '0 10px 0 10px',
+	// 	'maxHeight': '9vh',
+	// 	'minHeight': 72,
+	// 	'overflowY': 'hidden',
+	// 	'minWidth': '25.2vh',
+	// 	'margin': '10px',
+	// 	'marginTop': 8,
+	// 	'alignItems': 'center',
+	// 	'borderRadius': '8px',
+	// 	'backgroundColor': '#EDF2FB',
+	// 	'willChange': 'transform',
+	// 	'animation': 'mostClickedIdeas 20s linear infinite',
+	// 	'&:hover': {
+	// 		cursor: 'pointer',
+	// 	},
+	// },
+	// mostClickedIdeasTitle: {
+	// 	color: '#8197ae',
+	// 	fontSize: '.9rem',
+	// },
+	// mostClickedIdeasContent: {
+	// 	color: '#151515',
+	// 	fontSize: '1rem',
+	// },
 	carousel: {
 		borderRadius: '8px',
 		backgroundColor: '#fff',
 		border: '2px solid #EDEEF1',
 		padding: '10px 10px 2px 10px',
 		minHeight: 111,
-	},
-	carouselContect: {
-		'maxHeight': '4vh',
-		'textAlign': 'center',
-		'fontSize': '1.1rem',
-		'padding': '0 3% 0 3%',
-		'marginTop': 10,
-		'marginBottom': 10,
-		'&:hover': {
-			cursor: 'pointer',
-		},
 	},
 	column: {
 		'height': '100vh',
